requirements.txt
# MoltBook Watchdog - Dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # For parquet support

# Machine Learning
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.4

# Visualization
plotly>=5.18.0

# HTTP requests
requests>=2.31.0

# Progress bars
tqdm>=4.66.0
